requirements.txt
PyYAML==6.0.2
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
pyyaml>=6.0.0
jsonschema>=4.20.0